pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-flex"
version = "0.1.0"
description = "A cyberpunk-themed CLI dashboard for developers."
authors = [{name = "HuyLe0913", email = "huyphongtung@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich==14.1.0",
    "distro==1.9.0",
    "psutil==7.1.0",
    "typer==0.21.1",
    "speedtest-cli==2.1.3"
]

[project.scripts]
dev-flex = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
